Add new autoRealization policy, which allows the user to configure the behavior of the DataRealizer thread.  Now, by default, only global data is automatically realized.
Tweak closeDatafile and removeValue routines so the DataRepository doesn't
insist on realizing all the data in the file before closing it! Coupled with
the lazy autoRealization policy above, this dramatically speeds up the
action of renaming data.

Alter the DataNotifier so it catches recursive data function definitions and
prevents them from throwing the dashboard into infinite loops.  Also added
more comments to the DataNotifier so I could figure out what it was doing.

Add a check to avoid NullPointerExceptions when a datafile contains a line
of the form "foo=null"