PSP3 cycles need their own datafile.
Disabling (for now) the PSP3 cycle development script. You can get to this
script from the top-level script, so it isn't completely gone.  You just
aren't taken to it by default when you are in a cycle.  The reason for this
is that when viewing the cycle script, the project ID will be the ID for the
cycle, and that will prohibit effectively viewing top-level PSP3 data.