Added a patch for deletion logic - previously deleting PSP phases was allowed! Now nodes can only be deleted if they are editable or movable, or if their status field contains the "DELETE_OK_CHAR" (`D').
If a template's top-level node is not editable, then it cannot be added to
a node more than once.  This applies to insertable process phases like
"Design Review" (inserted into a PSP0 process).  We don't want people to be
able to insert it multiple times, yielding "Design Review", "Design Review1",
etc.