Enhancements to #if: Add new "defined" operator.  Also add "!" as an alternative to "not" operator.
Allow the option of parsing the text inserted by the #include directive.

Add a new "magic" variable, called "_UNIQUE_".  This variable can be
used in an #echo directive, and it will return a unique number every
time it is evaluated.  Web pages can incorporate the resulting number
in a URL query string to guarantee that the browser will not use a
cached object.