Major DataRepository refactor to reduce memory consumption.
Now, instead of mounting all data in the repository, the repository
will only immediately mount data that explicitly appears in a user's
datafile.  All of the definitions inherited from the #included file
are created lazily when they are first asked for.  When those lazy
elements are no longer in use, they are disposed.

This allows very full dashboard instances to start up and operate with
a memory footprint similar to an empty dashboard.  Startup times are
also improved.  Lazy creation of data happens quickly enough that
operating performance is not generally affected.

With the help of profiling tools, several other poorly performing
hotspots in the code were streamlined to avoid performance hits from
the new approach.  For example:

   * Clients requesting a list of repository keys can optionally
     supply "hints" about keys they are not interested in, to avoid
     unnecessary calls to createDataName().

   * The previous FrozenDataSet implementation has been replaced with
     a new, lightweight implementation, which does not use the
     RepositoryListener mechanism.  Instead, it assumes that data
     values are appropriately frozen or thawed at startup, and waits
     until a freeze flag actually changes value before doing any work.

   * The mechanism for disposing FormDataSession objects was broken,
     and has been fixed.

   * The export mechanism has been enhanced to perform fewer 
     java.util.regexp.Pattern operations, as they were using a
     significant percentage of the export time.