Update the TimingMetricsRecorder so it does not recalculate timing metrics while datafile changes are underway.
This was previously causing problems during the saving of Hierarchy
Editor renames.  The hierarchy would be in the middle of saving a 
list of changes to the data repository (for example, moving datafiles
from point A to point B), and would not yet have reached the final 
moves, but the metrics recorder would take it upon itself to fix the
timing mismatches too early.  Then, when the files finally did get
moved, data would be out of sync with the time log again.