When two packages with the same package ID exist in the search path, refine the algorithm used to select the best package.
Previously, this selection was based upon version numbers only, and would
leave multiple identical packages in the search path if they all had the
exact same version number.

Now, if two packages with the same ID and version are found, prefer the one
that is installed locally (in the same directory as pspdash.jar) rather than
one that was found in the template search path.  (This improves performance,
particularly over sluggish network connections.)