Started to implement ticket 128. Here are some notes about what has been made and what is left :
EVTaskList, mergedModel changed to AbstractTreeTableModel :
Necessary in order to get access to addTreeModelWillChangeListener.

EVTaskListRollup, added the hierarchy, the cache and the willNeedChangeNotification members :
Used by actionPerformed() to get a fresh copy of the saved task list.

EVTaskListRollup.actionPerformed() :
This is the only thing that does not work. This method is called when a contained task list is saved and it should reload the task list from scratch. There must be some event not being propagated correctly to the TaskScheduleDialog since it does not refresh properly. After this method is called, the only way for the TaskScheduleDialog to show the correct data for a task is to click on it (or mouse over it). Only then will it refresh.

EVTaskListRollup.treeStructureChanged() :
The call to recalc() must be inside the "if" block or else an infinite loop will occur when trying to open a rollup that contains other rollups. This is the method for which we tried to recalcTimer.restart() instead of the call to recalc() and outside the "if" block . It worked for a day, then stopped working for some unknown reason...

EVTaskListXML.getDataNameForTaskListName() :
This method has been added so that openXML() could access the dataName and then add the task list as a dataListener to the data repository.