When monitoring is enabled (to determine whether the lock has been lost), do not include a native release/relock step as part of each check. If the file is still in place and still contains the lock data we wrote earlier, then assume that we still own the lock.
This change is designed to avoid spurious errors.  It seems that releasing
and immediately re-acquiring the lock can fail some percentage of the time.
The LockWatcher was performing this operation once a minute for the entire
time an application is open.  This high-frequency operation, coupled with
the mean-time-between-failures of the release/relock operation, would
translate into unacceptably frequent error messages, telling the end-user
that "their lock was lost, and another user has claimed the lock in the
meantime."  These messages would commonly occur when all data was present
on the local hard drive (meaning that it is impossible for another user to
be editing the data anyway).

With this modification, we are no longer testing for the rare scenario
where the native lock has truly been lost.  This could occur if someone
leaves the application running while they disconnect from the network for
an extended period of time.  Previously, when they reconnect to the
network, the native lock would be re-acquired within a minute.  Now, the
lock data itself will be validated on reconnect, but the native lock might
not be re-acquired until they save data.  During the interim period,
another user might possibly steal the lock.

However, this change is the best compromise, because the "lock stolen"
logic will only kick in when the lock has truly been stolen by another user
(and such theft should be quite rare).  The previous behavior would
generate far too many false positives (telling the user that the lock had
been stolen when it had not).