Redacted backup files that were saved with Process Dashboard 1.15 - 1.15.7 do not contain projDump.xml files, because the old redaction logic discarded them.  That causes problems for the database-driven calculation logic, which requires the projDump.xml file to be present.
To avoid those problems, the quick launcher will now check for the presence of
wbs.xml and projDump.xml files when opening data backups.  If wbs files are
found but dump files are not, this is the hallmark of a redacted backup that
was saved with PD 1.15.  In that case, disable the database plugin and allow
rollups to be computed using the old calculation engine.