Bugfix for testing of Mac OS X client
Several older blocks of code were testing the os.name system property
for the value "Mac OS X", but the word "Mac" is no longer present in
newer versions of Mac Java. This change fixes the code blocks to perform
a more up-to-date test.

This change was implemented Jul 6, 2018 for Process Dashboard 2.4.2.3